+++ src/container/types.ts
@@ -15,6 +15,7 @@ import { ShortUrlDetail } from '../visits/reducers/shortUrlDetail';
 import { ShortUrlVisits } from '../visits/reducers/shortUrlVisits';
 import { TagVisits } from '../visits/reducers/tagVisits';
 import { DomainsList } from '../domains/reducers/domainsList';
+import { VisitsOverview } from '../visits/reducers/visitsOverview';
 
 export interface ShlinkState {
   servers: ServersMap;
@@ -35,6 +36,7 @@ export interface ShlinkState {
   mercureInfo: MercureInfo;
   settings: Settings;
   domainsList: DomainsList;
+  visitsOverview: VisitsOverview;
 }
 
 export type ConnectDecorator = (props: string[] | null, actions?: string[]) => any;